+++ tests/testthat/test-dilplots-ggplot.R
@@ -44,10 +44,10 @@ test_that("Able to plot good dilution data with ggplot correctly", {
     unique() %>%
     as.character()
 
-  dil_batch_col <- c("#377eb8")
+  curv_batch_col <- c("#377eb8")
 
   # Create palette for each dilution batch for plotting
-  pal <- dil_batch_col %>%
+  pal <- curv_batch_col %>%
     stats::setNames(dilution_batch_name)
 
   # Create dilution statistical summary